+++ src/Geopilot.Frontend/src/components/appSettings/appSettingsContext.tsx
@@ -2,6 +2,7 @@ import { createContext, FC, PropsWithChildren, useEffect, useState } from "react
 import { AppSettingsContextInterface, ClientSettings } from "./appSettingsInterface";
 import { useApi } from "../../api";
 import { ContentType } from "../../api/apiInterfaces.ts";
+import { useTranslation } from "react-i18next";
 
 export const AppSettingsContext = createContext<AppSettingsContextInterface>({
   initialized: false,
@@ -10,7 +11,8 @@ export const AppSettingsContext = createContext<AppSettingsContextInterface>({
 });
 
 export const AppSettingsProvider: FC<PropsWithChildren> = ({ children }) => {
-  const { fetchApi } = useApi();
+  const { i18n } = useTranslation();
+  const { fetchApi, fetchLocalizedMarkdown } = useApi();
   const [clientSettings, setClientSettings] = useState<ClientSettings | null>();
   const [termsOfUse, setTermsOfUse] = useState<string | null>();
 
@@ -18,10 +20,9 @@ export const AppSettingsProvider: FC<PropsWithChildren> = ({ children }) => {
     fetchApi<ClientSettings>("/client-settings.json", { responseType: ContentType.Json })
       .then(setClientSettings)
       .catch(() => setClientSettings(null));
-    fetchApi<string>("/terms-of-use.md", { responseType: ContentType.Markdown })
-      .then(setTermsOfUse)
-      .catch(() => setTermsOfUse(null));
-  }, [fetchApi]);
+
+    fetchLocalizedMarkdown("terms-of-use", i18n.language).then(setTermsOfUse);
+  }, [fetchApi, fetchLocalizedMarkdown, i18n.language]);
 
   useEffect(() => {
     if (clientSettings) {
@@ -61,8 +62,8 @@ export const AppSettingsProvider: FC<PropsWithChildren> = ({ children }) => {
     <AppSettingsContext.Provider
       value={{
         initialized: clientSettings !== undefined && termsOfUse !== undefined,
-        clientSettings: clientSettings,
         termsOfUse: termsOfUse,
+        clientSettings: clientSettings,
       }}>
       {children}
     </AppSettingsContext.Provider>